Building c8yctrl into main cumulocity-cypress package.
Refactoring of build and packaging. Moved dependencies into the main package. Improved and fixed imports.
