+++ package.json
@@ -1,25 +1,27 @@
 {
   "name": "cumulocity-cypress",
-  "version": "0.5.2",
+  "version": "0.7.5",
   "description": "Cypress commands for Cumulocity IoT",
   "scripts": {
-    "start-server": "cross-env NODE_NO_WARNINGS=1 npx tsx test/backend/test-server.ts",
+    "start:cypress:server": "cross-env NODE_NO_WARNINGS=1 npx tsx test/backend/test-server.ts",
+    "start:c8yctrl": "cd packages/pact-http-controller && npm run start",
     "test": "npm run jest:run && npm run cypress",
-    "cypress": "cross-env TZ=Europe/Berlin start-server-and-test start-server http://localhost:8080 cypress:run",
+    "cypress": "cross-env TZ=Europe/Berlin start-server-and-test start:cypress:server http://localhost:8080 cypress:run",
     "cypress:run": "cross-env TZ=Europe/Berlin cypress run --project test --browser chrome --config video=false",
-    "cypress:open": "cross-env TZ=Europe/Berlin start-server-and-test start-server http://localhost:8080 cypress:debug",
+    "cypress:open": "cross-env TZ=Europe/Berlin start-server-and-test start:cypress:server http://localhost:8080 cypress:debug",
     "cypress:debug": "cross-env CYPRESS_REMOTE_DEBUGGING_PORT=9222 cypress open --project test",
     "jest": "npm run jest:run",
     "jest:run": "cross-env NODE_NO_WARNINGS=1 jest --config jest.config.js",
     "clean": "rimraf dist/ && rimraf packages/*/dist/ && rimraf packages/*/.yalc && rimraf packages/*/yalc.lock",
-    "copy-files": "copyfiles --up 1 -V './src/lib/**/*.js' './src/lib/**/*.d.ts' dist/ && copyfiles '*.md' package.json dist/",
-    "build": "npm run clean && npm run copy-files && tsc -b -v src/ && npm run build:plugin",
-    "build:plugin": "rollup -c rollup.config.mjs",
-    "package": "npm run build && npm run clean:package:folder && cd dist/ && npm run clean:package:json",
+    "copy:lib": "copyfiles --up 1 -V './src/lib/**/*.js' './src/lib/**/*.d.ts' ./dist/ && copyfiles '*.md' package.json ./dist/",
+    "build": "npm run clean && npm run copy:lib && tsc -b -v src/ && npm run build:c8yctrl && npm run build:rollup",
+    "build:c8yctrl": "cd packages/pact-http-controller && npm run build",
+    "build:rollup": "rollup -c rollup.config.mjs",
+    "package": "npm run build && npm run clean:package:folder && cd ./dist/ && npm run clean:package:json",
     "yalc:runner": "npm run yalc:publish && cd packages/pact-runner && yalc add cumulocity-cypress",
     "yalc:publish": "npm run clean && npm run build && cd dist/ && yalc publish && cd ..",
     "yalc:controller": "npm run yalc:publish && cd packages/pact-http-controller && yalc add cumulocity-cypress",
-    "clean:package:folder": "rimraf -v -g './dist/**/*.map' && rimraf -v -g './dist/**/*.tsbuildinfo'",
+    "clean:package:folder": "rimraf -v -g './dist/**/*.tsbuildinfo'",
     "clean:package:json": "npm pkg delete 'devDependencies' && npm pkg delete 'scripts'",
     "lint": "eslint ./",
     "release": "semantic-release",
@@ -39,6 +41,9 @@
   "license": "Apache-2.0",
   "bugs": {
     "url": "https://github.com/SoftwareAG/cumulocity-cypress/issues"
+  },
+  "bin": {
+    "c8yctrl": "bin/c8yctrl.js"
   },
   "workspaces": [
     "packages/*"
@@ -60,14 +65,27 @@
       "types": "./lib/commands/c8ypact.d.ts",
       "default": "./lib/commands/c8ypact.js"
     },
+    "./c8yctrl": {
+      "types": "./shared/c8yctrl/index.d.ts",
+      "default": "./shared/c8yctrl/index.js"
+    },
+    "./node": {
+      "types": "./node.d.ts",
+      "default": "./node.js"
+    },
     "./commands/*": "./lib/commands/*",
-    "./c8ypact/*": "./lib/pact/*",
-    "./*": "./*"
+    "./c8ypact/*": "./lib/pact/*"
   },
   "typesVersions": {
     "*": {
       "c8ypact": [
         "./lib/commands/c8ypact.d.ts"
+      ],
+      "c8yctrl": [
+        "./c8yctrl.d.ts"
+      ],
+      "node": [
+        "./node.d.ts"
       ],
       "commands": [
         "./lib/commands/index.d.ts"
@@ -83,10 +101,23 @@
   "dependencies": {
     "ajv": "^8.12.0",
     "ajv-formats": "^2.1.1",
+    "body-parser": "^1.20.2",
+    "cookie": "^0.6.0",
+    "cookie-parser": "^1.4.6",
+    "cosmiconfig": "^9.0.0",
+    "cosmiconfig-typescript-loader": "^5.0.0",
     "date-fns": "^2.30.0",
+    "dotenv": "^16.4.5",
+    "express": "^4.18.2",
+    "http-proxy-middleware": "^3.0.0",
     "lodash": "^4.17.21",
+    "morgan": "^1.10.0",
     "quicktype": "^23.0.81",
-    "semver": "^7.5.2"
+    "semver": "^7.5.2",
+    "set-cookie-parser": "^2.6.0",
+    "tslib": "^2.6.2",
+    "winston": "^3.13.0",
+    "yargs": "^17.7.2"
   },
   "devDependencies": {
     "@rollup/plugin-alias": "^5.1.0",
@@ -98,6 +129,7 @@
     "@semantic-release/git": "^10.0.1",
     "@semantic-release/github": "^10.1.3",
     "@types/chai": "^4.3.4",
+    "@types/cookie": "^0.6.0",
     "@types/cookie-parser": "^1.4.6",
     "@types/debug": "^4.1.12",
     "@types/express": "^4.17.21",
@@ -122,6 +154,7 @@
     "rimraf": "^5.0.1",
     "rollup": "^4.14.1",
     "rollup-plugin-dts": "^6.1.0",
+    "rollup-plugin-includepaths": "^0.2.4",
     "rollup-plugin-node-resolve": "^5.2.0",
     "semantic-release": "^24.0.0",
     "semver": "^7.6.3",
