+++ rollup.config.mjs
@@ -1,7 +1,37 @@
 import dts from "rollup-plugin-dts";
 import resolve from "@rollup/plugin-node-resolve";
 import commonjs from "@rollup/plugin-commonjs";
+import alias from "@rollup/plugin-alias";
 import json from "@rollup/plugin-json";
+import path, { dirname } from "path";
+import { fileURLToPath } from "url";
+import includePaths from "rollup-plugin-includepaths";
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
+
+const ccPath = path.resolve(
+  __dirname,
+  "./packages/pact-http-controller/dist/src"
+);
+
+console.log("Resolving cumulocity-cypress to", ccPath);
+
+let includePathOptions = {
+  include: {},
+  paths: ["src"],
+  external: [],
+  extensions: [".js"],
+};
+
+const aliasConfig = {
+  entries: [
+    {
+      find: "cumulocity-cypress",
+      replacement: ccPath,
+    },
+  ],
+};
 
 export default [
   {
@@ -11,7 +41,7 @@ export default [
         name: "c8y",
         file: "dist/plugin/index.js",
         format: "umd",
-        sourcemap: true,
+        sourcemap: false,
       },
     ],
     plugins: [
@@ -24,7 +54,32 @@ export default [
   },
   {
     input: "dist/plugin/index.d.ts",
-    output: [{ file: "dist/plugin/index.d.ts", format: "es", sourcemap: true }],
+    output: [
+      { file: "dist/plugin/index.d.ts", format: "es", sourcemap: false },
+    ],
     plugins: [dts()],
+  },
+  {
+    input:
+      "./packages/pact-http-controller/dist/packages/pact-http-controller/src/startup.js",
+    output: [
+      {
+        name: "c8yctrl",
+        file: "dist/bin/c8yctrl.js",
+        format: "cjs",
+      },
+    ],
+    plugins: [
+      includePaths(includePathOptions),
+      resolve({
+        extensions: [".js"],
+        preferBuiltins: true,
+        mainFields: ["main", "module"],
+        resolveOnly: ["./packages/pact-http-controller/dist/**"],
+      }),
+      alias(aliasConfig),
+      commonjs(),
+      json(),
+    ],
   },
 ];
