+++ src/shared/c8yctrl/index.ts
@@ -1,0 +1,3 @@
+export * from "./httpcontroller";
+export * from "./httpcontroller-options";
+export * from "./../../shared/c8ypact";
