+++ tsconfig.json
@@ -5,7 +5,8 @@
     "module": "es2020",
     "moduleResolution": "node",
     "declaration": true,
-    "declarationMap": true,
+    "declarationMap": false,
+    "sourceMap": false,
     "strict": true,
     "skipLibCheck": true,
     "esModuleInterop": true,
